chore: fix bug in package.json, split up AddresGeocoder.tsx into sub-components
